+++ src/app/routes/productsrouter.ts
@@ -1,0 +1,5 @@
+import { Router } from 'express'
+
+export const productsRouter = Router()
+
+productsRouter.get('/products')
