--- src/app/services/Category/CategoryDetails.ts
@@ -1,2 +1,0 @@
-export * from './CreateCategoryService'
-export * from './ListCategoriesService'
