+++ src/app/services/Category/index.ts
@@ -1,3 +1,3 @@
-export * from './CategoryDetails'
+
 export * from './CreateCategoryService'
 export * from './ListCategoriesService'
