+++ src/app/usecases/Categories/ListCategories.ts
@@ -1,6 +1,12 @@
 import { Types } from 'mongoose'
-export interface ListCategories {
+export interface PersistListCategories {
   name: string
   icon: string
   _id: Types.ObjectId
+}
+
+export interface DomainListCategories {
+  name: string
+  icon: string
+  id: Types.ObjectId
 }
